@@ -20,7 +20,7 @@ version: 1.0.50+50
 
 environment:
   sdk: '>=3.4.4 <4.0.0'
-  flutter: 3.22.3
+  flutter: 3.24.0
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
